requirements.txt
streamlit
Pillow
pydrive